requirements.txt
Flask==3.1.1
fpdf==1.7.2
gunicorn==23.0.0
openai==1.96.1
python-dotenv==1.1.1
Flask-Bcrypt==1.0.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
Flask-Login===0.6.3
Flask-Migrate==4.1.0
email_validator==2.2.0
pydub==0.25.1
tiktoken==0.9.0